SupplyChainSimulation is a package to model and simulate supply chains.

The simulation is built around a loop that keeps track of orders and inventory movements:

receive_inventory

place_orders

receive_orders

send_inventory

Each step can be customized by applying different policies. The policies can also be optimized to improve the supply chain performance.

Installation

SupplyChainSimulation can be installed using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run

pkg> add SupplyChainSimulation

Getting Started

API

SupplyChainSimulation.eoq_cost_rate — Method

eoq_cost_rate(demand_rate, ordering_cost, holding_cost_rate)

Computes the total cost per time period of ordering the economic ordering quantity.

See also [`eoq_quantity`](@ref).

SupplyChainSimulation.eoq_interval — Method

eoq_interval(demand_rate, ordering_cost, holding_cost_rate)

Computes at what interval the economic ordering quantity is ordered.

See also [`eoq_quantity`](@ref).

SupplyChainSimulation.eoq_quantity — Method

eoq_quantity(demand_rate, ordering_cost, holding_cost_rate, backlog_cost_rate)

Computes the economic ordering quantity that minimizes overall costs (ordering costs + holding costs) while meeting  demand.

SupplyChainSimulation.eoq_quantity — Method

eoq_quantity(demand_rate, ordering_cost, holding_cost_rate)

Computes the economic ordering quantity that minimizes overall costs (ordering costs + holding costs) while meeting  demand.

SupplyChainSimulation.get_locations — Method

get_locations(network)

Gets all the locations in the network.

SupplyChainSimulation.FixedOrderingPolicy — Type

Orders the same amount; no matter the scenario.

SupplyChainSimulation.NetSSOrderingPolicy — Type

Orders up to a given number based on the net number of units (on hand + in transit + on order - on backlog) if the net inventory is below a threshold.

SupplyChainSimulation.NetUptoOrderingPolicy — Type

Orders up to a given number based on the net number of units (on hand + in transit + on order - on backlog).

SupplyChainSimulation.OnHandUptoOrderingPolicy — Type

Orders up to a given number based on the number of units on hand; no matter what is on order.

SupplyChainSimulation.QuantityOrderingPolicy — Type

Orders a given quantity specific to each time period.